refactor: NotificationStatus 패키지 이동
