+++ backend/bom-bom-server/src/test/java/me/bombom/api/v1/challenge/service/ChallengeStartNotificationServiceTest.java
@@ -7,8 +7,8 @@ import java.time.LocalDate;
 import java.util.List;
 import me.bombom.api.v1.TestFixture;
 import me.bombom.api.v1.challenge.domain.Challenge;
-import me.bombom.api.v1.challenge.domain.NotificationStatus;
 import me.bombom.api.v1.challenge.domain.notification.ChallengeStartNotification;
+import me.bombom.api.v1.challenge.domain.notification.NotificationStatus;
 import me.bombom.api.v1.challenge.repository.ChallengeParticipantRepository;
 import me.bombom.api.v1.challenge.repository.ChallengeRepository;
 import me.bombom.api.v1.challenge.repository.ChallengeStartNotificationRepository;
